requirements.txt
asgiref==3.5.2
# backports.zoneinfo==0.2.1
Django==4.1.3
django-extensions==3.2.1
djangorestframework==3.14.0
psycopg2-binary==2.9.5
pytz==2022.6
sqlparse==0.4.3
pillow==9.3.0
flake8==6.0.0
python-decouple==3.7
drf-yasg==1.21.4
matplotlib==3.6.3
